feat: add update note functionality and global error handler

+++ src/services/notes.service.ts
@@ -1,5 +1,6 @@
-import prisma from "../prismaClient/prismaClient.js";
+import prisma from "../../core/databaseClient/prismaClient/prismaClient.js";
 import type { Note } from "../../generated/prisma/client.js";
+import { NotFoundExcpetion } from "../../core/errors/NotFoundExcpetion.js";
 
 
 export function getNotes() {
@@ -14,10 +15,34 @@ export async function createNote(note: Note) {
 }
 
 export async function deleteNote(noteId: string) {
+    const note = await prisma.note.findUnique({
+        where: {
+            id: noteId
+        }
+    })
+    if (!note) throw new NotFoundExcpetion('Note not found');
+
+
     const deletedNote = await prisma.note.delete({
         where: {
             id: noteId
         }
     });
     return deletedNote;
+}
+
+export async function updateNote(noteId: string, data: Note) {
+    const updatedNote = await prisma.note.update({
+        where: {
+            id: noteId
+        },
+        data: {
+            title: data.title,
+            content: data.content,
+            privacy: data.privacy,
+            favorite: data.favorite,
+            status: data.status
+        }
+    });
+    return updatedNote;
 }
